weltladen/requirements.txt
-i https://pypi.python.org/simple
babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
django-admin-sortable2==0.7.6
django-allauth==0.41.0
django-angular==2.2.4
django-appconf==1.0.4
django-classy-tags==1.0.0
django-cms==3.7.1
django-entangled==0.3
django-filer==1.7.0
django-filter==2.2.0
django-formtools==2.2
django-fsm-admin==1.2.4
django-fsm==2.7.0
django-haystack==2.8.1
django-ipware==2.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-parler-rest==2.1
django-parler==2.0.1
django-phonenumber-field==4.0.0
django-polymorphic==2.1.2
django-post-office==3.3.1
django-redis-cache==2.1.0
django-redis-sessions==0.6.1
django-rest-auth==0.9.5
django-sass-processor==0.8
django-sekizai==1.1.0
django-select2==7.2.3
django-shop==1.1.1
django-treebeard==4.3.1
django==2.2.12
djangocms-admin-style==1.5.0
djangocms-bootstrap==1.1.2
djangocms-cascade==1.2.3
djangocms-text-ckeditor==3.9.0
djangorestframework==3.10.3
drf-haystack==1.8.6
easy-thumbnails==2.7
elasticsearch==1.7.0
html5lib==1.0.1
idna==2.9
jsonfield==3.1.0
libsass==0.19.4
oauthlib==3.1.0
phonenumbers==8.12.1
pillow==7.1.0
psycopg2-binary==2.7.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests-oauthlib==1.3.0
requests==2.23.0
schedule==0.6.0
six==1.14.0
sqlparse==0.3.1
unidecode==1.1.1
urllib3==1.25.8
uwsgi==2.0.18
webencodings==0.5.1
